fix wsgen fork=true invocation
